backend/requirements.txt
Django==4.2.*
djangorestframework
django-cors-headers==4.4.0
google-generativeai
python-dotenv
httpx
markdown
feedparser==6.0.11